changed: more dense packing of type info pointers
This comes at a tradeoff where it is no longer O(1) to look up the information
about a type from the `type_expr`. It is now O(log n) in order to perform the
binary search required. This is likely preferable, but it might be worth
changing the algorithm to something that has better *practical* timings,
such as a linear-search.